+++ containers-tests/tests/intmap-strictness.hs
@@ -8,8 +8,9 @@ import Data.Bifunctor (bimap)
 import Data.Coerce (coerce)
 import Data.Either (partitionEithers)
 import qualified Data.Foldable as F
-import Data.Functor.Identity (Identity(..))
 import Data.Function (on)
+import Data.Functor.Compose
+import Data.Functor.Identity (Identity(..))
 import qualified Data.List as List
 import qualified Data.List.NonEmpty as NE
 import Data.Maybe (catMaybes, mapMaybe)
@@ -17,6 +18,7 @@ import Data.Ord (comparing)
 import Test.ChasingBottoms.IsBottom
 import Test.Tasty (TestTree, defaultMain, testGroup)
 import Test.Tasty.QuickCheck (testProperty)
+import Data.Tuple.Solo (Solo (MkSolo), getSolo)
 import Test.QuickCheck
 import Test.QuickCheck.Poly (A, B, C)
 import Test.QuickCheck.Function (apply)
@@ -68,12 +70,35 @@ prop_strictFromSet :: Func Key (Bot A) -> IntSet -> Property
 prop_strictFromSet fun set =
   isBottom (M.fromSet f set) === any (isBottom . f) (IntSet.toList set)
   where
-    f = coerce (applyFunc fun) :: Key -> A
+    f = applyFunc fun
+
+prop_strictFromSetA :: Func Key (Bot A) -> IntSet -> Property
+prop_strictFromSetA fun set =
+  isBottom (getSolo (M.fromSetA (MkSolo . f) set)) === any (isBottom . f) (IntSet.toList set)
+  where
+    f = applyFunc fun
 
 prop_lazyFromSet :: Func Key (Bot A) -> IntSet -> Property
 prop_lazyFromSet fun set = isNotBottomProp (L.fromSet f set)
   where
-    f = coerce (applyFunc fun) :: Key -> A
+    f = applyFunc fun
+
+prop_lazyFromSetA :: Func Key (Bot A) -> IntSet -> Property
+prop_lazyFromSetA fun set = isNotBottomProp (getSolo (L.fromSetA f set))
+  where
+  f = MkSolo . applyFunc fun
+
+prop_fromSetA_equiv_strictness :: Func Int (Bot A) -> IntSet -> Property
+prop_fromSetA_equiv_strictness fun set =
+  -- strict fromSetA is the same as lazy and then forcing
+  bottomOn (M.fromSetA f set) (fmap forceValues (L.fromSetA f set)) .&&.
+  -- strict fromSetA is the same as lazy fromSetA composed with strictly applied
+  -- wrapper
+  bottomOn (M.fromSetA f set) (fmap getSolo . getCompose $ L.fromSetA (Compose . fmap (MkSolo $!) . f) set)
+  where
+  forceValues xs = foldr (\ !_ r -> r) () xs `seq` xs
+  bottomOn = (===) `on` isBottom . getSolo
+  f = MkSolo . applyFunc fun
 
 prop_strictFromList :: [(Key, Bot A)] -> Property
 prop_strictFromList kvs =
@@ -1015,6 +1040,8 @@ tests =
     , testGroup "Construction"
       [ testPropStrictLazy "singleton" prop_strictSingleton prop_lazySingleton
       , testPropStrictLazy "fromSet" prop_strictFromSet prop_lazyFromSet
+      , testPropStrictLazy "fromSetA" prop_strictFromSetA prop_lazyFromSetA
+      , testProperty       "fromSetA equivalences" prop_fromSetA_equiv_strictness
       , testPropStrictLazy "fromList" prop_strictFromList prop_lazyFromList
       , testPropStrictLazy "fromListWith" prop_strictFromListWith prop_lazyFromListWith
       , testPropStrictLazy "fromListWithKey" prop_strictFromListWithKey prop_lazyFromListWithKey