+++ containers/src/Data/Map/Strict/Internal.hs
@@ -214,6 +214,7 @@ module Data.Map.Strict.Internal
     , keysSet
     , argSet
     , fromSet
+    , fromSetA
     , fromArgSet
 
     -- ** Lists
@@ -420,9 +421,7 @@ import Utils.Containers.Internal.StrictPair
 import Data.Coerce
 #endif
 
-#ifdef __GLASGOW_HASKELL__
 import Data.Functor.Identity (Identity (..))
-#endif
 
 import qualified Data.Foldable as Foldable
 
@@ -1455,8 +1454,43 @@ mapKeysWith c f m =
 -- > fromSet undefined Data.Set.empty == empty
 
 fromSet :: (k -> a) -> Set.Set k -> Map k a
-fromSet _ Set.Tip = Tip
-fromSet f (Set.Bin sz x l r) = case f x of v -> v `seq` Bin sz x v (fromSet f l) (fromSet f r)
+#ifdef __GLASGOW_HASKELL__
+fromSet f = runIdentity . fromSetA (coerce f)
+#else
+fromSet f = runIdentity . fromSetA (pure . f)
+#endif
+
+-- | \(O(n)\). Build a map from a set of keys and a function which for each key
+-- computes its value in an 'Applicative' context.
+--
+-- This can only be as strict as the 'Applicative' allows it to be.
+--
+-- > fromSetA (\k -> pure $ replicate k 'a') (Data.Set.fromList [3, 5]) == pure (fromList [(5,"aaaaa"), (3,"aaa")])
+-- > fromSetA undefined Data.Set.empty == pure empty
+--
+-- The following strictness properties hold:
+--
+-- > fromSetA f = fmap forceValues . Data.Map.Lazy.fromSetA f
+-- >   where
+-- >     forceValues xs = foldr (\ !_ r -> r) () xs `seq` xs
+--
+-- > fromSetA f =
+-- >   fmap getSolo .
+-- >   getCompose .
+-- >   Data.Map.Lazy.fromSetA (Compose . fmap (MkSolo $!) . f)
+
+fromSetA :: Applicative f => (k -> f a) -> Set.Set k -> f (Map k a)
+fromSetA _ Set.Tip = pure Tip
+fromSetA f (Set.Bin sz x l r) = 
+  flip (Bin sz x $!)
+    <$> fromSetA f l
+    <*> f x
+    <*> fromSetA f r
+#if __GLASGOW_HASKELL__
+{-# INLINABLE fromSetA #-}
+#else
+{-# INLINE fromSetA #-}
+#endif
 
 -- | \(O(n)\). Build a map from a set of elements contained inside 'Arg's.
 --